Refactor RdfService into some util classes containing static function and another service called GraphService.
