+++ backend/Domain/Services/IRdfService.cs
@@ -1,8 +1,6 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using OpenData.API.Domain.Models;
-using OpenData.API.Domain.Models.Queries;
-using OpenData.API.Domain.Services.Communication;
 using System;
 using VDS.RDF;
 using VDS.RDF.Writing;
