+++ backend/Services/GraphService.cs
@@ -1,0 +1,213 @@
+using System;
+using VDS.RDF;
+using VDS.RDF.Writing;
+using VDS.RDF.Parsing;
+using System.Threading.Tasks;
+using OpenData.API.Domain.Models;
+using OpenData.API.Domain.Services;
+using OpenData.API.Domain.Repositories;
+using System.Collections.Generic;
+using System.Linq;
+using Newtonsoft.Json.Linq;
+using OpenData.API.Util;
+
+
+namespace OpenData.API.Services
+{
+    public class GraphService : IGraphService
+    {
+        private readonly IDatasetRepository _datasetRepository;
+        private readonly IDistributionRepository _distributionRepository;
+        private readonly IPublisherRepository _publisherRepository;
+        private readonly ICategoryRepository _categoryRepository;
+        private readonly ITagsRepository _tagsRepository;
+        private readonly IUnitOfWork _unitOfWork;
+
+        public GraphService(IDatasetRepository datasetRepository, IDistributionRepository distributionRepository, IPublisherRepository publisherRepository, ICategoryRepository categoryRepository, ITagsRepository tagsRepository, IUnitOfWork unitOfWork)
+        {
+            _datasetRepository = datasetRepository;
+            _distributionRepository = distributionRepository;
+            _publisherRepository = publisherRepository;
+            _categoryRepository = categoryRepository;
+            _tagsRepository = tagsRepository;
+            _unitOfWork = unitOfWork;
+        }
+
+        // Add a dataset in a graph to the database
+        public async Task<Dataset> AddDataset(Graph g) 
+        {
+            // Find publisher id
+            Publisher publisher = await AddPublisher(g);
+
+            // Find the dataset subject uri 
+            IUriNode dcatDataset = g.CreateUriNode("dcat:Dataset");
+            String[] datasetUri = findSubjectUri(g, dcatDataset).Split(",");
+            // From the dataset uri make a dictionary with the attributes
+            Dictionary<string,string> attributes = getAttributesFromSubject(g, datasetUri[0]);
+            // Add relevant attributes to a new dataset
+            Dataset dataset = new Dataset {
+                Title = attributes.GetValueOrDefault("title", "Ingen tittel"), 
+                Identifier = datasetUri[0], 
+                Description = attributes.GetValueOrDefault("description", ""), 
+                PublicationStatus = attributes.ContainsKey("distribution") ? EPublicationStatus.published : EPublicationStatus.notPublished,
+                PublisherId = publisher.Id, 
+                CategoryId = 100 
+            };
+
+            // Add the dataset to the database
+            await _datasetRepository.AddAsync(dataset);
+            await _unitOfWork.CompleteAsync();
+            
+            AddTags(g, attributes.GetValueOrDefault("keyword", ""), dataset);
+            await AddDistribution(g, dataset.Id);
+            return dataset;
+        }
+
+        // Add a distribution in a graph to the database
+        public async Task<List<Distribution>> AddDistribution(Graph g, int datasetId)
+        {
+            // Find the distribution subject uri 
+            IUriNode dcatDistribution = g.CreateUriNode("dcat:Distribution");
+            String[] distributionUris = findSubjectUri(g, dcatDistribution).Split(",");
+            List<Distribution> distributions = new List<Distribution>();
+            foreach (String distributionUri in distributionUris)
+            {
+                Dictionary<string,string> attributes = getAttributesFromSubject(g, distributionUri);
+                //Parse file format
+                EFileFormat fileFormat = EFileFormat.annet;
+                String[] fileFormatString = attributes.GetValueOrDefault("format", "annet").Split(",");
+                try 
+                {
+                    fileFormat = (EFileFormat)Enum.Parse(typeof(EFileFormat), fileFormatString[0], true);
+                }
+                catch(Exception ex){Console.WriteLine(ex.Message);}
+
+                // Add relevant attributes to a new distribution
+                Distribution distribution = new Distribution {
+                    Title = (string) attributes.GetValueOrDefault("title", attributes.GetValueOrDefault("description", "")), 
+                    Uri = (string) attributes.GetValueOrDefault("accessURL", ""), 
+                    FileFormat = fileFormat, 
+                    DatasetId = datasetId
+                };
+
+                // Add the dataset to the distribution
+                await _distributionRepository.AddAsync(distribution);
+                await _unitOfWork.CompleteAsync();
+                distributions.Add(distribution);
+            }
+            return distributions;
+        }
+
+        // Add a publisher in a graph to the database
+        public async Task<Publisher> AddPublisher(Graph g)
+        {
+            // Find the publisher subject uri 
+            IUriNode foafOrganization = g.CreateUriNode("foaf:Organization");
+            String[] publisherUri = findSubjectUri(g, foafOrganization).Split(",");
+            if (String.IsNullOrWhiteSpace(publisherUri[0]))
+            {
+                IUriNode foafAgent = g.CreateUriNode("foaf:Agent");
+                publisherUri = findSubjectUri(g, foafAgent).Split(",");
+            }
+            
+            Dictionary<string,string> attributes = getAttributesFromSubject(g, publisherUri[0]);
+            String publisherName = attributes.GetValueOrDefault("name", "Ukjent");
+            
+            // Check if the publisher  already exists
+            IEnumerable<Publisher> existingPublishers = await _publisherRepository.ListAsync();
+            foreach (Publisher pub in existingPublishers)
+            {
+                if (pub.Name.ToLower().Equals(publisherName.ToLower()))
+                {
+                    return pub;
+                }
+            }
+
+            // If it doesn't exist, add relevant attributes to a new publisher
+            Publisher publisher = new Publisher {
+                Name = publisherName, 
+            };
+
+            // Add the dataset to the publisher
+            await _publisherRepository.AddAsync(publisher);
+            await _unitOfWork.CompleteAsync();
+            return publisher;
+        }
+
+        public async void AddTags(Graph g, String keywords, Dataset dataset)
+        {
+            String[] keywordsList = keywords.Split(",");
+            IEnumerable<Tags> existingTags = await _tagsRepository.ListAsync();
+            foreach (String keyword in keywordsList)
+            {
+                Tags existingTag = null;
+                foreach (Tags tag in existingTags) 
+                {
+                    if (tag.Name.ToLower().Equals(keyword.ToLower()))
+                    {
+                        existingTag = tag;
+                        break;
+                    }
+                }
+                if (existingTag != null) 
+                {
+                    DatasetTags datasetTag = new DatasetTags { Dataset = dataset, Tags = existingTag };
+                    dataset.DatasetTags.Add(datasetTag);
+                }
+                else 
+                {
+                    Tags tag = new Tags { Name = keyword };
+                    await _tagsRepository.AddAsync(tag);
+                    await _unitOfWork.CompleteAsync();
+                    DatasetTags datasetTag = new DatasetTags { Dataset = dataset, Tags = tag };
+                    dataset.DatasetTags.Add(datasetTag);
+                    await _unitOfWork.CompleteAsync();
+                }
+            }
+        }
+
+        // Get attributes as dictionary from chosen subject node
+        private Dictionary<string,string> getAttributesFromSubject(Graph g, String subject)
+        {
+            Dictionary<string,string> attributes = new Dictionary<string, string>();
+            IEnumerable<Triple> triples = g.Triples;
+            foreach (Triple t in triples){
+                if (t.Subject.ToString() == subject)
+                {
+                    string p = t.Predicate.ToString();
+                    int slash = p.LastIndexOf("/");
+                    int hash = p.LastIndexOf("#");
+                    int index = Math.Max(slash,hash) + 1;
+                    p = p.Substring(index);
+                    if(attributes.ContainsKey(p))
+                    {
+                        attributes[p] = attributes[p] + "," + t.Object.ToString();
+                    }
+                    else
+                    {
+                        attributes[p] = t.Object.ToString();
+                    }
+                }
+            }
+            return attributes;
+        }
+        // Find the subject uri for a chosen object uriNode using the predicate rdf:type
+        private String findSubjectUri(Graph g, IUriNode uriNode)
+        {
+            IUriNode rdfType = g.CreateUriNode("rdf:type");
+            String uri = "";
+            IEnumerable<Triple> ts = g.GetTriplesWithPredicateObject(rdfType, uriNode);
+            foreach (Triple t in ts) {
+                if (String.IsNullOrEmpty(uri))
+                {
+                    uri = t.Subject.ToString();
+                }
+                else 
+                {
+                    uri += "," + t.Subject.ToString();
+                }
+            }
+            return uri;
+        }
+    }
+}
